shipment creation in the page logic
