+++ src/app/src/routes/(map)/+layout.svelte
@@ -1,8 +1,6 @@
 <script lang="ts">
 	import { defaultLocation } from '$lib/common';
 	import { MapLibre } from 'svelte-maplibre';
-	import Header from '$components/common/Header.svelte';
-	import '../../styles.scss';
 	import type { Snippet } from 'svelte';
 
 	const { children } = $props<{
@@ -10,15 +8,11 @@
 	}>();
 </script>
 
-<main>
-	<Header />
-
-	<MapLibre
-		style="https://basemaps.cartocdn.com/gl/positron-gl-style/style.json"
-		class="maplibre relative z-0 h-screen w-full sm:aspect-video"
-		zoom={15}
-		center={defaultLocation}
-	>
-		{@render children()}
-	</MapLibre>
-</main>
+<MapLibre
+	style="https://basemaps.cartocdn.com/gl/positron-gl-style/style.json"
+	class="maplibre relative z-0 h-screen w-full sm:aspect-video"
+	zoom={15}
+	center={defaultLocation}
+>
+	{@render children()}
+</MapLibre>
