+++ src/app/src/stories/create-shipment.stories.ts
@@ -1,22 +1,24 @@
-import type { Meta, StoryObj } from '@storybook/svelte';
-import CreateShipment from '$components/forms/CreateShipment.svelte';
-import type { SvelteComponent } from 'svelte';
+// import type { Meta, StoryObj } from '@storybook/svelte';
+// import CreateShipment from '$components/forms/CreateShipment.svelte';
+// import type { SvelteComponent } from 'svelte';
 
-const meta = {
-	title: 'Forms/CreateShipment',
-	component: CreateShipment,
-	tags: ['autodocs']
-} satisfies Meta<SvelteComponent>;
+// const meta = {
+// 	title: 'Forms/CreateShipment',
+// 	component: CreateShipment,
+// 	tags: ['autodocs']
+// } satisfies Meta<SvelteComponent>;
 
-export default meta;
-type Story = StoryObj<typeof meta>;
+// export default meta;
+// type Story = StoryObj<typeof meta>;
 
-// Basic story showing the create shipment form
-export const Basic: Story = {
-	args: {
-		showModal: true,
-		onClose: () => {
-			console.log('closeHandler');
-		}
-	}
-};
+// // Basic story showing the create shipment form
+// export const Basic: Story = {
+// 	args: {
+// 		showModal: true,
+// 		onClose: () => {
+// 			console.log('closeHandler');
+// 		}
+// 	}
+// };
+
+// TODO: migrate to page stories
